+++ SourceCode/xl/String/xlString.h
@@ -242,7 +242,7 @@ namespace xl
     template <typename T>
     inline int StringT<T>::Length() const
     {
-        return m_aData.Size() - 1;
+        return (int)m_aData.Size() - 1;
     }
 
     template <typename T>
@@ -370,7 +370,7 @@ namespace xl
     template <typename T>
     inline StringT<T> StringT<T>::Right(int nLength) const
     {
-        return SubString(m_aData.Size() - 1 - nLength, nLength);
+        return SubString((int)m_aData.Size() - 1 - nLength, nLength);
     }
 
     template <typename T>
